Update dependencies
Update dependency build process to the use the CMAKE define of
`BUILD_DEPS` to trigger the process of updating and building the
dependencies for the source.  This processm atches that of
VulkanTools and Validation Layer repos so that it's easier to
automate and incorporate into our build process in the future.

Also update the dependencies to a more recent version (not several
years old).