cdl: Implement VK_NV_device_diagnostic_checkpoints
This extension provides more reliable and simpler command buffer
tracking than VK_AMD_buffer_marker. But it is much less flexible.

In particular, it cannot be used implement semaphore tracking
because it doesn't provide any way to find an order of events
outside of a single queue. So, restructure the code so that
command buffer and queue progress can use either extension
via the Checkpoint class. But semaphore tracker uses the
Marker class which is an abstraction of VK_AMD_buffer_marker only.