command: Improve state management
Rather than have GetCommandBufferState() look at checkpoints,
move this to a separate method called UpdateStateFromCheckpoints().
This is called at the last possible moment after we know there's
been a crash. Otherwise the checkpoint values are somewhat
unreliable.

Rename command buffer and command states to match what it is printed
in the dump file so that its easier to go back and forth.